requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0
requests==2.32.3
